requirements-local.txt
flask
mlflow
pandas
google-cloud-storage
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-storage
feast
scikit-learn==1.5.2
xgboost==2.1.3
numpy
matplotlib
db-dtypes
